requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
mysql-connector-python
pydantic[email]
python-dotenv

